feat(primer): refractor + fix sleep + change constants + define speeds + add plots
animals were sleeping if they happened to accidentally go to the edge of the world from wandering,
this was fixed so that an animal only sleeps if its previous objective was to sleep. fixed position
for home and animals will sleep not when its "in range" but when it has reached it. made all food to
be the same size as there was no difference between larger and smaller food. changed constants for
analysis in paper. added data collection and plots for average age, food eaten/left (not working
yet) and births/deaths. PrimerData now also saves start params in json.