feat(microbe model): introduce fixes and improvements to the microbe model
fixed some mistakes in the implementation of the model. added new data to be collected, saved and
plotted for analysis. also made small change to simulation_controller to only schedule canvas update
if it is not updating at the moment. analyse_microbe.py was used for analysis of specific files